requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
Babel==2.8.0
colorama==0.4.3
coverage==5.2.1
Django==3.1
django-cpf==0.1.0
django-nose==1.4.6
django-phonenumber-field==4.0.0
django-rest-framework==0.1.0
django-safedelete==0.5.6
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.91
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
model-bakery==1.1.1
more-itertools==8.4.0
nose==1.3.7
packaging==20.4
phonenumbers==8.12.7
pinocchio==0.4.2
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
